+++ src/cmd/main.cpp
@@ -353,7 +353,7 @@ int main(int argc, char *argv[]) {
         if(filename != "unknown") {
             // filename given, try to get signature
             sig = ELFGetSignature(filename.c_str());
-            if(sig != -1) {
+            if(sig != (unsigned int)-1) {
                 // signature in elf found, try to get devicename
                 std::map<unsigned int, std::string>::iterator cur  = AvrSignatureToNameMap.find(sig);
                 if(cur != AvrSignatureToNameMap.end()) {
